feat: complete componentization admin, cart page
