+++ src/basic/components/ui/CouponForm.tsx
@@ -1,0 +1,99 @@
+import { FocusEvent } from 'react'
+import { Coupon } from '../../../types'
+
+export function CouponForm({
+  couponForm,
+  handleCouponSubmit,
+  setShowCouponForm,
+  handleEditCouponForm,
+  handleDiscountValueValidation,
+}: {
+  couponForm: Coupon
+  setShowCouponForm: React.Dispatch<React.SetStateAction<boolean>>
+  handleCouponSubmit: (e: React.FormEvent) => void
+  handleEditCouponForm: (
+    e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>,
+    key: string,
+  ) => void
+  handleDiscountValueValidation: (e: FocusEvent<HTMLInputElement>) => void
+}) {
+  return (
+    <div className="mt-6 p-4 bg-gray-50 rounded-lg">
+      <form onSubmit={handleCouponSubmit} className="space-y-4">
+        <h3 className="text-md font-medium text-gray-900">새 쿠폰 생성</h3>
+        <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-4">
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              쿠폰명
+            </label>
+            <input
+              type="text"
+              value={couponForm.name}
+              onChange={(e) => handleEditCouponForm(e, 'name')}
+              className="w-full border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 px-3 py-2 border text-sm"
+              placeholder="신규 가입 쿠폰"
+              required
+            />
+          </div>
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              쿠폰 코드
+            </label>
+            <input
+              type="text"
+              value={couponForm.code}
+              onChange={(e) => handleEditCouponForm(e, 'code')}
+              className="w-full border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 px-3 py-2 border text-sm font-mono"
+              placeholder="WELCOME2024"
+              required
+            />
+          </div>
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              할인 타입
+            </label>
+            <select
+              value={couponForm.discountType}
+              onChange={(e) => handleEditCouponForm(e, 'discountType')}
+              className="w-full border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 px-3 py-2 border text-sm"
+            >
+              <option value="amount">정액 할인</option>
+              <option value="percentage">정률 할인</option>
+            </select>
+          </div>
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              {couponForm.discountType === 'amount' ? '할인 금액' : '할인율(%)'}
+            </label>
+            <input
+              type="text"
+              value={
+                couponForm.discountValue === 0 ? '' : couponForm.discountValue
+              }
+              onChange={(e) => handleEditCouponForm(e, 'discountValue')}
+              onBlur={handleDiscountValueValidation}
+              className="w-full border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 px-3 py-2 border text-sm"
+              placeholder={couponForm.discountType === 'amount' ? '5000' : '10'}
+              required
+            />
+          </div>
+        </div>
+        <div className="flex justify-end gap-3">
+          <button
+            type="button"
+            onClick={() => setShowCouponForm(false)}
+            className="px-4 py-2 border border-gray-300 rounded-md text-sm font-medium text-gray-700 hover:bg-gray-50"
+          >
+            취소
+          </button>
+          <button
+            type="submit"
+            className="px-4 py-2 bg-indigo-600 text-white rounded-md text-sm font-medium hover:bg-indigo-700"
+          >
+            쿠폰 생성
+          </button>
+        </div>
+      </form>
+    </div>
+  )
+}
