+++ src/types.ts
@@ -1,24 +1,28 @@
 export interface Product {
-  id: string;
-  name: string;
-  price: number;
-  stock: number;
-  discounts: Discount[];
+  id: string
+  name: string
+  price: number
+  stock: number
+  discounts: Discount[]
+}
+
+export interface ProductForm extends Omit<Product, 'id'> {
+  description: string
 }
 
 export interface Discount {
-  quantity: number;
-  rate: number;
+  quantity: number
+  rate: number
 }
 
 export interface CartItem {
-  product: Product;
-  quantity: number;
+  product: Product
+  quantity: number
 }
 
 export interface Coupon {
-  name: string;
-  code: string;
-  discountType: 'amount' | 'percentage';
-  discountValue: number;
+  name: string
+  code: string
+  discountType: 'amount' | 'percentage'
+  discountValue: number
 }
